requirement.txt
pytesseract
pyautogui
pillow